@@ -13,15 +13,15 @@
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "^5.0.3",
-        "@angular/common": "^5.0.3",
-        "@angular/compiler": "^5.0.3",
-        "@angular/core": "^5.0.3",
-        "@angular/forms": "^5.0.3",
-        "@angular/http": "^5.0.3",
-        "@angular/platform-browser": "^5.0.3",
-        "@angular/platform-browser-dynamic": "^5.0.3",
-        "@angular/router": "^5.0.3",
+        "@angular/animations": "^5.2.5",
+        "@angular/common": "^5.2.5",
+        "@angular/compiler": "^5.2.5",
+        "@angular/core": "^5.2.5",
+        "@angular/forms": "^5.2.5",
+        "@angular/http": "^5.2.5",
+        "@angular/platform-browser": "^5.2.5",
+        "@angular/platform-browser-dynamic": "^5.2.5",
+        "@angular/router": "^5.2.5",
         "@aspnet/signalr": "1.0.0-preview1-28189",
         "@types/bootstrap": "^3.3.33",
         "@types/jquery": "^3.2.12",
@@ -32,7 +32,7 @@
         "@types/moment-timezone": "^0.2.34",
         "@types/signalr": "^2.2.33",
         "@types/toastr": "^2.1.33",
-        "abp-ng2-module": "^1.3.0",
+        "abp-ng2-module": "^2.0.0",
         "abp-web-resources": "^3.3.0",
         "animate.css": "^3.5.2",
         "block-ui": "^2.70.1",
@@ -69,8 +69,8 @@
         "zone.js": "0.8.18"
     },
     "devDependencies": {
-        "@angular/cli": "^1.5.4",
-        "@angular/compiler-cli": "^5.0.3",
+        "@angular/cli": "^1.6.8",
+        "@angular/compiler-cli": "^5.2.5",
         "@angularclass/hmr": "^2.1.3",
         "@types/jasmine": "^2.5.38",
         "@types/node": "^8.0.27",
