@@ -1,4 +1,6 @@
-﻿import { NgModule } from '@angular/core';
+import { NgModule } from '@angular/core';
+import { HTTP_INTERCEPTORS } from '@angular/common/http';
+import { AbpHttpInterceptor } from '@abp/abpHttpInterceptor';
 
 import * as ApiServiceProxies from './service-proxies';
 
@@ -10,7 +12,8 @@ import * as ApiServiceProxies from './service-proxies';
         ApiServiceProxies.UserServiceProxy,
         ApiServiceProxies.TokenAuthServiceProxy,
         ApiServiceProxies.AccountServiceProxy,
-        ApiServiceProxies.ConfigurationServiceProxy
+        ApiServiceProxies.ConfigurationServiceProxy,
+        { provide: HTTP_INTERCEPTORS, useClass: AbpHttpInterceptor, multi: true }
     ]
 })
 export class ServiceProxyModule { }
